Refactor: Move spatial indexing to farm.core.spatial
This commit reorganizes the spatial indexing components into a dedicated `farm.core.spatial` package. The original `spatial_index.py` now acts as a compatibility shim, re-exporting the classes from the new module. This improves code organization and maintainability.
